Featur/checkbox

+++ src/main/java/com/buildflow/erp/domain/achats/dto/response/LigneAchatResponse.java
@@ -9,7 +9,7 @@ public record LigneAchatResponse(
         String designation,
         BigDecimal quantite,
         String unite,
-        BigDecimal prixUnitaire,
+        double prixUnitaire,
         BigDecimal total,
         String bpuLigneRef
 ) {}
